Fixed potential drive issue.
The robot was not driving because it was checking if it was following the frisbee in the #drive(move, turn) method, which is called by both drive with joystick (which needs to be blocked if following frisbee) and by the frisbee following (which can't be blocked when following frisbee). Fixed by moving check to #driveWithJoystick(joytick) method.